+++ scripts/macros/2014/classFeatures/illrigger/interdictBoons/unleashHell.js
@@ -1,0 +1,29 @@
+import {itemUtils} from '../../../../../utils.js';
+async function added({trigger: {entity: item}}) {
+    await itemUtils.multiCorrectActivityItemConsumption(item, ['use'], {
+        0: 'balefulInterdict',
+        1: 'interdictBoons'
+    });
+}
+export let interdictBoonUnleashHell = {
+    name: 'Interdict Boons: Unleash Hell',
+    version: '1.3.77',
+    rules: 'legacy',
+    item: [
+        {
+            pass: 'created',
+            macro: added,
+            priority: 50
+        },
+        {
+            pass: 'itemMedkit',
+            macro: added,
+            priority: 50
+        },
+        {
+            pass: 'actorMunch',
+            macro: added,
+            priority: 50
+        }
+    ]
+};