+++ scripts/macros/2024/classFeatures/rogue/steadyAim.js
@@ -14,7 +14,7 @@ async function use({trigger, workflow}) {
 }
 export let steadyAim = {
     name: 'Steady Aim',
-    version: '1.3.53',
+    version: '1.3.77',
     rules: 'modern',
     midi: {
         item: [